+++ BoostRunClub/Services/RunningService/RunningMotionServiecable.swift
@@ -1,15 +1,14 @@
 //
-//  MotionActivityProvidable.swift
+//  RunningMotionServicable.swift
 //  BoostRunClub
 //
 //  Created by 김신우 on 2020/12/20.
 //
 
 import Combine
-import CoreMotion
 import Foundation
 
-protocol MotionActivityProvidable {
+protocol RunningMotionServiceable {
     var motionTypeSubject: PassthroughSubject<MotionType, Never> { get }
     func start()
     func stop()
