+++ BoostRunClub/Services/RunningService/RunningRecordService.swift
@@ -9,16 +9,6 @@ import Combine
 import CoreLocation
 import Foundation
 
-protocol RunningRecordServiceable {
-    func addState(_ state: RunningState)
-    func save(startTime: Date?, endTime: Date?) -> (activity: Activity, detail: ActivityDetail)?
-    func clear()
-
-    var locations: [CLLocation] { get }
-    var routes: [RunningSlice] { get }
-    var didAddSplitSignal: PassthroughSubject<RunningSplit, Never> { get }
-}
-
 final class RunningRecordService: RunningRecordServiceable {
     private var cancellables = Set<AnyCancellable>()
     private let activityWriter: ActivityWritable
